Move qpu_available to tests/__init__.py
Since it is only used by the code under tests
